requirements.txt
# BurnSight Requirements
# Tested on Google Colab (Python 3.10+)

tensorflow>=2.13.0
tensorflow-probability>=0.21.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.7.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
scikit-image>=0.20.0
scipy>=1.10.0
h5py>=3.8.0
cupy-cuda12x>=12.0.0       # For CUDA 12.x; change to cupy-cuda11x if using CUDA 11.x
albumentations>=1.3.0
imageio>=2.28.0
joblib>=1.2.0